FMOD cleanup
Now THIS is a lot cleaner.  It also fixes the bug of the sounds cutting out after a few seconds (basically FMOD ceasing to play sounds) as I was not using the update function for each frame or tic as per the documentation demands.  Now the fmod audio system is updated inside I_StartDisplay which is called each new tic or frame and it is all good.