beginnings of 3D positional audio
What this will do is calculate the distance between origin and current position (when it is finished).  So that sounds that originated far away will then playback quieter or not at all.  So if you opened a door and ran through maybe 10 metres, then the next sound closing the door should be very quiet.  Same with monster roaming sounds.