Making audio tracks format agnostic
FMOD does not require the audio extension in the filename as it reads the files and finds the format.  This way, modders don't need to add .ogg or .mp3 and instead will just drop the file without an extension into the directory and can mix and match whatever supported formats they desire without any consequences.